+++ src/MentorLake.Gtk3/GLib/GMainContextFlags.cs
@@ -1,7 +1,7 @@
 namespace MentorLake.GLib;
 
 [Flags]
-public enum GMainContextFlags : long
+public enum GMainContextFlags : uint
 {
 	G_MAIN_CONTEXT_FLAGS_NONE = 0,
 	G_MAIN_CONTEXT_FLAGS_OWNERLESS_POLLING = 1
