+++ src/MentorLake.Gtk3/Gio/GDBusCapabilityFlags.cs
@@ -1,7 +1,7 @@
 namespace MentorLake.Gio;
 
 [Flags]
-public enum GDBusCapabilityFlags : long
+public enum GDBusCapabilityFlags : uint
 {
 	G_DBUS_CAPABILITY_FLAGS_NONE = 0,
 	G_DBUS_CAPABILITY_FLAGS_UNIX_FD_PASSING = 1
