+++ src/MentorLake.Gtk3/Gio/GDBusInterfaceSkeletonFlags.cs
@@ -1,7 +1,7 @@
 namespace MentorLake.Gio;
 
 [Flags]
-public enum GDBusInterfaceSkeletonFlags : long
+public enum GDBusInterfaceSkeletonFlags : uint
 {
 	G_DBUS_INTERFACE_SKELETON_FLAGS_NONE = 0,
 	G_DBUS_INTERFACE_SKELETON_FLAGS_HANDLE_METHOD_INVOCATIONS_IN_THREAD = 1
