+++ src/MentorLake.Gtk3/Gio/GDBusObjectManagerClientFlags.cs
@@ -1,7 +1,7 @@
 namespace MentorLake.Gio;
 
 [Flags]
-public enum GDBusObjectManagerClientFlags : long
+public enum GDBusObjectManagerClientFlags : uint
 {
 	G_DBUS_OBJECT_MANAGER_CLIENT_FLAGS_NONE = 0,
 	G_DBUS_OBJECT_MANAGER_CLIENT_FLAGS_DO_NOT_AUTO_START = 1
