+++ src/MentorLake.Gtk3/Gio/GDBusSendMessageFlags.cs
@@ -1,7 +1,7 @@
 namespace MentorLake.Gio;
 
 [Flags]
-public enum GDBusSendMessageFlags : long
+public enum GDBusSendMessageFlags : uint
 {
 	G_DBUS_SEND_MESSAGE_FLAGS_NONE = 0,
 	G_DBUS_SEND_MESSAGE_FLAGS_PRESERVE_SERIAL = 1
