+++ src/MentorLake.Gtk3/Gio/GFileCreateFlags.cs
@@ -1,7 +1,7 @@
 namespace MentorLake.Gio;
 
 [Flags]
-public enum GFileCreateFlags : long
+public enum GFileCreateFlags : uint
 {
 	G_FILE_CREATE_NONE = 0,
 	G_FILE_CREATE_PRIVATE = 1,
