+++ src/MentorLake.Gtk3/Gio/GIOStreamSpliceFlags.cs
@@ -1,7 +1,7 @@
 namespace MentorLake.Gio;
 
 [Flags]
-public enum GIOStreamSpliceFlags : long
+public enum GIOStreamSpliceFlags : uint
 {
 	G_IO_STREAM_SPLICE_NONE = 0,
 	G_IO_STREAM_SPLICE_CLOSE_STREAM1 = 1,
