+++ src/MentorLake.Gtk3/Gio/GMountUnmountFlags.cs
@@ -1,7 +1,7 @@
 namespace MentorLake.Gio;
 
 [Flags]
-public enum GMountUnmountFlags : long
+public enum GMountUnmountFlags : uint
 {
 	G_MOUNT_UNMOUNT_NONE = 0,
 	G_MOUNT_UNMOUNT_FORCE = 1
