+++ src/MentorLake.Gtk3/Gio/GOutputStreamSpliceFlags.cs
@@ -1,7 +1,7 @@
 namespace MentorLake.Gio;
 
 [Flags]
-public enum GOutputStreamSpliceFlags : long
+public enum GOutputStreamSpliceFlags : uint
 {
 	G_OUTPUT_STREAM_SPLICE_NONE = 0,
 	G_OUTPUT_STREAM_SPLICE_CLOSE_SOURCE = 1,
