+++ src/MentorLake.Gtk3/Gio/GSocketMsgFlags.cs
@@ -1,7 +1,7 @@
 namespace MentorLake.Gio;
 
 [Flags]
-public enum GSocketMsgFlags : long
+public enum GSocketMsgFlags : uint
 {
 	G_SOCKET_MSG_NONE = 0,
 	G_SOCKET_MSG_OOB = 1,
