+++ src/MentorLake.Gtk3/Gio/GTestDBusFlags.cs
@@ -1,7 +1,7 @@
 namespace MentorLake.Gio;
 
 [Flags]
-public enum GTestDBusFlags : long
+public enum GTestDBusFlags : uint
 {
 	G_TEST_DBUS_NONE = 0
 }
